git-svn-id: http://svn.codehaus.org/geoserver/branches/wcs-branch@3532 ef1d6e69-97e5-0310-af46-8a06194da32a
Former-commit-id: 62f52cb8611b750d325ab0d8c10c5524574bfd64 [formerly 37f2e2d631de1da9a94f73edeefef8b0db69f721]
Former-commit-id: 7ee607b1f0f7656c1d6ea8e9d59271d0fca55e76

+++ geoserver/src/org/vfny/geoserver/global/xml/XMLConfigWriter.java
@@ -1030,31 +1030,33 @@ public class XMLConfigWriter {
 		
 		File[] fa = dir.listFiles();
 		for(int j=0;j<fa.length;j++){
-			// find dir name
-			i = data.getCoverages().values().iterator();
-			CoverageInfoDTO cvi = null;
-			while(cvi==null && i.hasNext()){
-				CoverageInfoDTO cv = (CoverageInfoDTO)i.next();
-				if(cv.getDirName().equals(fa[j].getName())){
-					cvi = cv;
-				}
-			}
-			if(cvi == null){
-				//delete it
-				File[] t = fa[j].listFiles();
-				if (t != null) {
-					for(int x=0;x<t.length;x++) {
-						//hold on to the data, but be sure to get rid of the
-						//geoserver config shit, as these were deleted.
-						if (t[x].getName().equals("info.xml")) {
-							//sorry for the hardcodes, I don't remember if/where
-							//we have these file names.
-							t[x].delete();
-						}
+			if(fa[j].isDirectory()) {
+				// find dir name
+				i = data.getCoverages().values().iterator();
+				CoverageInfoDTO cvi = null;
+				while(cvi==null && i.hasNext()){
+					CoverageInfoDTO cv = (CoverageInfoDTO)i.next();
+					if(cv.getDirName().equals(fa[j].getName())){
+						cvi = cv;
 					}
 				}
-				if (fa[j].listFiles().length == 0) {
-					fa[j].delete();
+				if(cvi == null){
+					//delete it
+					File[] t = fa[j].listFiles();
+					if (t != null) {
+						for(int x=0;x<t.length;x++) {
+							//hold on to the data, but be sure to get rid of the
+							//geoserver config shit, as these were deleted.
+							if (t[x].getName().equals("info.xml")) {
+								//sorry for the hardcodes, I don't remember if/where
+								//we have these file names.
+								t[x].delete();
+							}
+						}
+					}
+					if (fa[j].listFiles().length == 0) {
+						fa[j].delete();
+					}
 				}
 			}
 		}
